requirements.txt
flask>=2.2,<4.0
python-pam>=2.0,<3.0
psutil>=5.8,<6.0
gpiozero>=1.6,<2.0
requests>=2.30,<3.0
werkzeug>=2.2,<4.0
six>=1.15,<2.0
gunicorn>=20.0,<21.0

